Fixing bug

+++ bankrollBust/pygameUtils/buttonUtils.py
@@ -245,5 +245,5 @@ class inputBox(button):
                     self.value = self.value + key_pressed # Adds the input to the value string
         if self.inputType == "num":
             # Returns the value as a integer if it isn't empty
-            return int(self.value) if self.value != "" else 0
+            return float(self.value) if self.value != "" else 0
         return self.value # returns the value string
